Refactored extensions and their tests into separate directories
